Use a workaround to fix bad pyfakefs behavior with pytest 7.0.0
It seems pyfakefs has a little problem when pytest is started like this:
  pytest tests/

When the `fs` argument is used in a test, the project path is transformed
into a pyfakefs object. So pytest cannot use the path correctly: assertion.

Luckily, we can change the current directory and run pytest correctly with:
  cd tests/ && pytest

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.fr>