Add kalray plugin to configure DPU
This plugin allows to create raid, logical volume store and logical
volumes on the Kalray DPU. It also allows the deletion of a volume.
Currently restrications are due to the Kalray DPU. Next generation of
the DPU will allow more volumes and won't be restricted on volume name.
See the README for details.

The unittest mocks the RPC server and it checks that required parameters
are well set and that all parameters used are valid.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>